+++ packages/axway-cli-oum/src/lib/activity.js
@@ -17,18 +17,28 @@ export async function renderActivity({ account, console, json, results }) {
 	const { createTable } = require('@axway/amplify-cli-utils');
 	const { default: snooplogg } = require('snooplogg');
 	const { highlight, note } = snooplogg.styles;
-	const { from, to, events } = results;
+	let { from, to, events } = results;
+
+	const formatDate = d => {
+		const dt = d ? new Date(d) : new Date();
+		return `${dt.getUTCMonth() + 1}/${dt.getUTCDate()}/${dt.getUTCFullYear()}`;
+	};
 
 	console.log(`Account:      ${highlight(account.name)}`);
 	if (results.org) {
 		console.log(`Organization: ${highlight(results.org.name)} ${note(`(${results.org.guid})`)}`);
 	}
-	console.log(`Date Range:   ${highlight(new Date(from).toLocaleDateString())} - ${highlight((to ? new Date(to) : new Date()).toLocaleDateString())}\n`);
+	console.log(`Date Range:   ${highlight(formatDate(from))} - ${highlight(formatDate(to))}\n`);
+
+	if (!events.length) {
+		console.log('No activity found');
+		return;
+	}
 
 	const table = createTable([ 'Date', 'Description', 'Event Name', 'Organization' ]);
 	for (const event of events) {
 		let changes = '';
-		if (Array.isArray(event.data.changes)) {
+		if (event.data && Array.isArray(event.data.changes)) {
 			const t = str => str.toLowerCase().replace(/(?:^|\s|-)\S/g, c => c.toUpperCase());
 			changes = event.data.changes.map((c, i, arr) => {
 				let unit;
@@ -60,22 +70,26 @@ export async function renderActivity({ account, console, json, results }) {
 				}
 
 				let desc = `${i + 1 === arr.length ? '└─' : '├─'} ${highlight(c.k)}`;
-				if (c.v !== undefined) {
-					if (c.o !== undefined) {
-						desc += ` changed from ${highlight(`"${c.o}"`)} to ${highlight(`"${c.v}"`)}`;
-					} else {
-						desc += ` of ${highlight(`"${c.v}"`)} was added`;
-					}
+
+				if (c.v === true || c.v === false) {
+					desc += ` was ${c.v ? 'enabled' : 'disabled'}`;
+				} else if (c.o !== undefined && c.v !== undefined) {
+					desc += ` changed from ${highlight(`"${c.o}"`)} to ${highlight(`"${c.v}"`)}`;
+				} else if (c.v !== undefined) {
+					desc += ` of ${highlight(`"${c.v}"`)} was added`;
+				} else {
+					desc += ` of ${highlight(`"${c.o}"`)} was removed`;
 				}
+
 				return desc;
 			}).join('\n');
 		}
 
 		table.push([
-			new Date(event.ts).toLocaleDateString(),
+			formatDate(event.ts),
 			event.message.replace(/__s__(.*?)__\/s__/g, (s, m) => highlight(m)),
 			event.event,
-			event.data.org_name || note('n/a')
+			event.data?.org_name || note('n/a')
 		]);
 
 		if (changes) {